Moved non-script files to lib subdirectory.
