+++ script/db_maintenance/lib/enhanced_connection.rb
