+++ script/db_maintenance/lib/postgres_helpers.rb
